Update tsconfig to add more strict rules.
This involves changing some existing files to ensure
that they don't break the build. Upon doing this, I exposed
possible bugs, and now fixed some of them and added
tests to guard for them in the future.

I notice that I use a lot of `<Something> | undefined` in my
functions. I don't know if I like this patern. I think it may be
cleaner if I jsut throw errors instead,.....

Hmmmm I have to reflect on this.