Add interface and class for transaction
I ended up spending a lot of time thinking about
how I should structure this. This is what I ended up
with. I am not sure if this is the best way to do it,
I will likely need to do some more reading about
SQL tables and what not to know how to do it best.