Implement inferAccountPath and add tests.
I know that the tests looks a little silly, since it just has the
exact same mapping in both the implementaion and the
test file.

I think this is okay for now since the implementation will
change in the future to query the accountPathMap from
another source, and the test should still remain correct.