Add tests for inferAccount
This is why you write tests..... I realized that the
function actually didn't work with the mapping we had stored.

This again may look dumb right now since the mapping
is currently static, and it's just the same in both files.
But the intention behind is to update it as soon as I figure
out where I can store the mapping.

These tests in place to ensure I don't break the functions
in the future.